skip already downloaded files

+++ scraper.py
@@ -32,6 +32,10 @@ async def download_images(url, folder):
             for index, image in enumerate(images):
                 src = image["src"]
                 filename = f"{index}.jpg"
+                # and args.overwrite:
+                if Path(f"{folder}/{filename}").exists():
+                    print("Skipped {filename}")
+                    continue
                 async with session.get(src) as response:
                     if response.ok:
                         with open(f"{folder}/{filename}", "wb") as f:
@@ -68,10 +72,10 @@ if __name__ == "__main__":
     parser.add_argument("url", help="URL to download images from")
     parser.add_argument("start_chapter", help="chapter to start downloading from")
     parser.add_argument("max_chapters", help="Maximum number of chapters to download")
-    parser.add_argument("folder", help="Folder to save images to", nargs="?")
-    args = parser.parse_args()
+    parser.add_argument("folder", help="Folder to save images to", nargs="?", default=Path(__file__).parent / "Manga")
+    #parser.add_argument("--overwrite", help="Overwrite old files", default=True)
 
-    outpath = Path(args.folder) if args.folder else Path(__file__).parent / "Manga"
+    args = parser.parse_args()
 
     HEADERS = {}
     HEADERS.update(get_referer(args.url))
@@ -79,8 +83,8 @@ if __name__ == "__main__":
     loop = asyncio.get_event_loop()
     tasks = []
     if r"{chapter}" not in args.url:
-        tasks.append(crawl(args.url, outpath, int(args.start_chapter), int(args.max_chapters)))
+        tasks.append(crawl(args.url, args.folder, int(args.start_chapter), int(args.max_chapters)))
     else:
         for i in range(int(args.start_chapter), int(args.max_chapters) + 1):
-            tasks.append(download_images(args.url.format(chapter=i), outpath / f"ch{i}"))
+            tasks.append(download_images(args.url.format(chapter=i), args.folder / f"ch{i}"))
     loop.run_until_complete(asyncio.wait(tasks))
